requirements.txt
apache-airflow==2.4.3
apache-airflow-providers-elasticsearch==4.4.0
apache-airflow-providers-amazon==8.7.0
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-mysql==5.0.0
apache-airflow-providers-mongo==3.2.2
apache-airflow-providers-google==10.9.0
apache-airflow-providers-common-sql
apache-airflow-providers-smtp==1.3.2
apache-airflow-providers-sftp==4.2.4
apache-airflow-providers-atlassian-jira==2.0.1
apache-airflow-providers-telegram==4.0.0
apache-airflow-providers-redis==3.1.0
pandas==1.5.3
numpy==1.21.6
boto3
pyarrow==19.0.0
python-dotenv~=0.20.0
pendulum~=2.1.2
selenium~=4.2.0
webdriver-manager==3.7.0
beautifulsoup4==4.12.2
openpyxl==3.0.10
requests
clickhouse-driver==0.2.4
airflow-clickhouse-plugin==1.4.0
kafka-python==2.0.2
fastavro==1.9.4
pydantic-avro
Unidecode==1.3.4
XlsxWriter==3.0.3
python-dateutil==2.8.2
pg8000==1.29.1
python-telegram-bot
gspread==4.0.1
SQLAlchemy==1.4.41
simplejson==3.17.6
redis
PyMySQL==1.0.2
pytz==2022.2.1
Pillow==9.3.0
confluent_avro==1.8.0
jira==3.4.1
pycryptodome==3.15.0
xlrd==2.0.1
gspread-dataframe==3.3.0
pyzipper==0.3.6
pysftp==0.2.9
python-gnupg==0.5.0
asyncio==3.4.3
asyncio-throttle==1.0.2
Jinja2
botocore
PyYAML
pydantic
psycopg2
cryptography==39.0.2
PyJWT
Deprecated
holidays==0.61
click>=8.1.7
inquirer>=3.4.0
clickhouse-connect>=0.8.9
hubspot-api-client==8.2.1
msoffcrypto-tool==5.4.2
pyOpenSSL==23.0.0
